Deploy all DevOps services for development
A deployment of a "koji" server consists of additional services that
support development beyond koji itself.  These include
git/cgit/gitolite, a mash containing the current snapshot of the
packages in koji, and an upstreams package sources cache.

Because these are required for integration with a DevOps workflow, they
are not included by default in the ansible playbook.