Make setup for koji NFS mounting easier
When koji builder and koji master are hosted on different machines, NFS
mounting is needed to connect the services.  Make this process more
smooth to minimize manual editing of files.  Namely, change the top
directory that koji looks at.  If on the master, this is on the local
file system.  If on the builder not on the master, this is also on the
local file system, but is backed by a NFS mount in a different location.

Signed-off-by: George T Kramer <george.t.kramer@intel.com>