Replace mash utility with bash script
The mash utility has not been ported to python3.  There is also no
comparable replacement utility of a similar scope that can do a similar
job.  There is pungi utility, which is the replacement for mash.
However, pungi is Fedora's build process encapsulated into a single tool
and is doing too much for simply creating repos for distribution.

To avoid this complexity, koji does have a dist-repo subcommand.
However, this takes too long to execute at around 15 minutes on powerful
hardware.

In lieu of not wanting to maintain mash any further, write a small bash
script that performs the same function using artifacts for koji on the
disk.

Signed-off-by: George T Kramer <george.t.kramer@intel.com>