Conditionally run swupd during setup
swupd should only be run during setup when on an officially released
version of Clear that has an upadate stream.  Development versions are
usually indicated when the VERSION_ID is equal to 1 and do not have an
update stream.  In these cases, the assumption is that all required
software is already installed prior to setup.

However, add checks for only the most critical dependencies before
continuing with the setup so that it more obvious to the user that these
are missing when they are.

Signed-off-by: George T Kramer <george.t.kramer@intel.com>