Supply more sane SSL certificate defaults
Want to enable spinning up a development environment quickly.

Increasing the countryName_max SSL configuration did not allow a country
name of 'Example Country Code' to be used when generating a certificate.

Prefer to revert back to the original value for countryName_max and
supply shorter parameters for certificate generation.

Change challengePassword_max to be consistent with the other limits.

Signed-off-by: George T Kramer <george.t.kramer@intel.com>